refactor: get ratio.js from ERTP
