+++ a3p-integration/proposals/z:acceptance/test/test-lib/vaults.js
@@ -10,10 +10,7 @@ import {
   GOV1ADDR,
   GOV2ADDR,
 } from '@agoric/synthetic-chain';
-import {
-  ceilMultiplyBy,
-  makeRatio,
-} from '@agoric/zoe/src/contractSupport/ratio.js';
+import { ceilMultiplyBy, makeRatio } from '@agoric/ertp/src/ratio.js';
 import { smartWalletKit } from './index.js';
 import { listVaults, vstorageKit } from './utils.js';
