+++ multichain-testing/scripts/fast-usdc-tool.ts
@@ -8,7 +8,7 @@ import type { ExecutionContext } from 'ava';
 import { encodeAddressHook } from '@agoric/cosmic-proto/address-hooks.js';
 import { AmountMath, type Brand } from '@agoric/ertp';
 import type { USDCProposalShapes } from '@agoric/fast-usdc/src/pool-share-math.js';
-import { divideBy } from '@agoric/zoe/src/contractSupport/ratio.js';
+import { divideBy } from '@agoric/ertp/src/ratio.js';
 import { makeDenomTools } from '../tools/asset-info.js';
 import { makeDoOffer } from '../tools/e2e-tools.js';
 import { commonSetup } from '../test/support.js';
