+++ packages/fast-usdc-contract/test/mocks.ts
@@ -6,7 +6,7 @@ import type {
   OrchestrationAccount,
 } from '@agoric/orchestration';
 import type { VowTools } from '@agoric/vow';
-import { makeRatio } from '@agoric/zoe/src/contractSupport/ratio.js';
+import { makeRatio } from '@agoric/ertp/src/ratio.js';
 import type { AmountUtils } from '@agoric/zoe/tools/test-utils.js';
 import type { Zone } from '@agoric/zone';
 import type { FeeConfig, LogFn } from '@agoric/fast-usdc/src/types.js';
