+++ packages/inter-protocol/src/interest-math.js
@@ -6,7 +6,7 @@ import {
   invertRatio,
   multiplyRatios,
   ratiosSame,
-} from '@agoric/zoe/src/contractSupport/ratio.js';
+} from '@agoric/ertp/src/ratio.js';
 
 /**
  * @import {Amount, Brand, Payment} from '@agoric/ertp';
