+++ packages/inter-protocol/src/vaultFactory/math.js
@@ -12,7 +12,7 @@ import {
   ceilMultiplyBy,
   floorDivideBy,
   ratioGTE,
-} from '@agoric/zoe/src/contractSupport/ratio.js';
+} from '@agoric/ertp/src/ratio.js';
 import { priceFrom } from '../auction/util.js';
 import { addSubtract } from '../contractSupport.js';
