+++ packages/inter-protocol/test/interest-math.test.js
@@ -1,7 +1,7 @@
 import { test } from '@agoric/zoe/tools/prepare-test-env-ava.js';
 
 import { makeIssuerKit, AmountMath } from '@agoric/ertp';
-import { makeRatio } from '@agoric/zoe/src/contractSupport/ratio.js';
+import { makeRatio } from '@agoric/ertp/src/ratio.js';
 import { calculateCurrentDebt, reverseInterest } from '../src/interest-math.js';
 
 const brand = makeIssuerKit('foo').brand;
