+++ packages/inter-protocol/test/vaultFactory/interestSupport.js
@@ -1,8 +1,5 @@
 import { AmountMath } from '@agoric/ertp';
-import {
-  floorMultiplyBy,
-  makeRatio,
-} from '@agoric/zoe/src/contractSupport/ratio.js';
+import { floorMultiplyBy, makeRatio } from '@agoric/ertp/src/ratio.js';
 import { Far } from '@endo/marshal';
 import { reverseInterest } from '../../src/interest-math.js';
