+++ packages/inter-protocol/test/vaultFactory/prioritizedVaults.test.js
@@ -1,10 +1,7 @@
 import { test } from '@agoric/zoe/tools/prepare-test-env-ava.js';
 
 import { AmountMath, makeIssuerKit } from '@agoric/ertp';
-import {
-  makeRatio,
-  makeRatioFromAmounts,
-} from '@agoric/zoe/src/contractSupport/ratio.js';
+import { makeRatio, makeRatioFromAmounts } from '@agoric/ertp/src/ratio.js';
 
 import { makeScalarBigMapStore } from '@agoric/vat-data';
 import {
