+++ packages/inter-protocol/test/vaultFactory/vault-contract-wrapper.js
@@ -15,7 +15,7 @@ import {
   makeRatio,
   multiplyBy,
   multiplyRatios,
-} from '@agoric/zoe/src/contractSupport/ratio.js';
+} from '@agoric/ertp/src/ratio.js';
 import { makeFakePriceAuthority } from '@agoric/zoe/tools/fakePriceAuthority.js';
 import { buildZoeManualTimer } from '@agoric/zoe/tools/manualTimer.js';
 import { E } from '@endo/eventual-send';
