+++ packages/zoe/test/unitTests/contracts/scaledPriceAuthority.test.js
@@ -8,7 +8,7 @@ import { E } from '@endo/eventual-send';
 import { makeIssuerKit, AmountMath } from '@agoric/ertp';
 
 import { makeZoeForTest } from '../../../tools/setup-zoe.js';
-import { makeRatio } from '../../../src/contractSupport/ratio.js';
+import { makeRatio } from '../../../src/contractSupport/index.js';
 import { makeFakeVatAdmin } from '../../../tools/fakeVatAdmin.js';
 import buildManualTimer from '../../../tools/manualTimer.js';
 import { makeManualPriceAuthority } from '../../../tools/manualPriceAuthority.js';
